requirements.txt
absl-py==0.15.0
addict==2.4.0
asttokens==2.2.1
astunparse==1.6.3
attrs==19.3.0
Automat==0.8.0
backcall==0.2.0
biwrap==0.1.6
black==21.12b0
blinker==1.4
cachetools==5.3.2
catkin-pkg==0.5.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==8.1.7
cloud-init==23.2.2
cloudpickle==2.2.1
cmake==3.26.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.1.1
cryptography==2.8
cycler==0.11.0
DateTime==5.1
dbus-python==1.2.16
decorator==4.4.2
distro==1.4.0
distro-info==0.23+ubuntu1.1
dm-tree==0.1.8
dnspython==1.16.0
docutils==0.20.1
entrypoints==0.3
executing==1.2.0
fail2ban==0.11.1
filelock==3.13.1
flake8==4.0.1
flatbuffers==1.12
fonttools==4.45.1
fsspec==2023.10.0
gast==0.3.3
google-auth==2.23.4
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.59.3
gssapi==1.6.1
gym==0.21.0
h5py==2.10.0
httplib2==0.14.0
huggingface-hub==0.19.4
hyperlink==19.0.0
idna==2.8
imageio==2.33.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.7.0
importlib-resources==6.1.1
incremental==16.10.1
iotop==0.6
ipaclient==4.8.6
ipalib==4.8.6
ipaplatform==4.8.6
ipapython==4.8.6
ipython_genutils==0.2.0
isc==2.0
jedi==0.18.2
Jinja2==2.10.1
joblib==1.3.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-core==4.6.3
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.4.4
kornia==0.5.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lit==16.0.6
llvmlite==0.39.1
lxml==4.5.0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.4
matplotlib-inline==0.1.6
mccabe==0.6.1
mlnx-tools===-5.2.0-
mmcv-full==1.7.1
more-itertools==4.2.0
mpmath==1.3.0
mypy-extensions==1.0.0
netaddr==0.7.19
netifaces==0.10.4
networkx==3.1
numba==0.56.0
numpy==1.22.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
nvsm==22.9.8
nvsm-dshm==22.9.8
nvsm-health==22.9.8
nvsm-sdk==1.0.0
oauthlib==3.1.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==23.2
paho-mqtt==1.5.0
pandas==1.4.3
parso==0.8.3
pathspec==0.11.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==4.0.0
ply==3.11
protobuf==3.19.6
psutil==5.5.1
pure-eval==0.2.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.8.0
pycparser==2.19
pycrypto==2.6.1
pyflakes==2.4.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.1+ubuntu0.20.4.1
python-augeas==0.5.0
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-ldap==3.2.0
python-yubico==1.3.3
pytz==2023.3.post1
pyusb==1.0.2
PyYAML==6.0
pyzmq==18.1.1
qrcode==6.1
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.9
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.10.1
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.5.6
ssh-import-id==5.10
stable-baselines3==1.6.0
stack-data==0.6.2
sympy==1.12
systemd-python==234
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-plot==0.3.2
tensorflow-probability==0.11.1
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
timm==0.9.12
tomli==1.2.3
torch==1.9.0
torchvision==0.10.0
tornado==5.1.1
tqdm==4.64.0
traitlets==5.9.0
triton==2.1.0
Twisted==18.9.0
typing_extensions==4.8.0
ubuntu-advantage-tools==8001
ufw==0.36
urllib3==1.25.8
wadllib==1.3.3
wcwidth==0.2.6
Werkzeug==3.0.1
yacs==0.1.8
yapf==0.40.2
zipp==3.17.0
zope.interface==4.7.1
